pyproject.toml
[project]
name = "tmp"
version = "0.1.0"
description = "FastAPI backend with LangChain"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "langchain",
    "openai",
    "pydantic",
    "python-dotenv",
    "sqlalchemy>=2.0.38",
    "passlib>=1.7.4",
    "jwt>=1.3.1",
    "pydantic-settings>=2.7.1",
    "ruff>=0.9.5",
    "bcrypt==4.0.1",
]
requires-python = "==3.13.2"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 20


#[tool.uv]
#module = "backend"
#
#[build-system]
#requires = ["uv==0.5.29"]
#buil-backend = "uv"
